requirements.txt
python-binance
python-dotenv
python-telegram-bot
